requirements.txt
# Web Framework
streamlit>=1.51.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client
requests>=2.32.0

# PDF Processing
pypdf>=4.0.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0

# Firebase
firebase-admin>=6.4.0
